requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
djangogrpcframework==0.2.1
djangorestframework==3.13.1
grpcio==1.48.1
grpcio-tools==1.48.1
protobuf==3.20.1
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
